+++ com.unity.render-pipelines.lightweight/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [5.10.0] - 2019-XX-XX
+### Added
+- Added RenderObjects. You can add RenderObjects to a Renderer to perform custom rendering.
+
+### Fixed
+- Fixed an XR rendering issue where Unity required a depth texture.
 
 ## [5.9.0] - 2019-03-15
 ### Fixed